add comments, refactor code
