+++ app/src/androidTest/kotlin/jp/co/yumemi/android/code_check/ui/view/PreviewScreenTest.kt
@@ -11,6 +11,7 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import jp.co.yumemi.android.code_check.model.Owner
 import jp.co.yumemi.android.code_check.model.RepositoryItem
+import jp.co.yumemi.android.code_check.ui.compose.InfoSection
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
