+++ app/src/androidTest/kotlin/jp/co/yumemi/android/code_check/ui/view/TopAppBarTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.ui.compose.GithubRepoAppTopAppBar
 import jp.co.yumemi.android.code_check.ui.theme.GithubRepositoryAppTheme
 import org.junit.Rule
 import org.junit.Test
