+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/compose/ErrorScreen.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.ui.view
+package jp.co.yumemi.android.code_check.ui.compose
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
