+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/compose/HomeScreen.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.ui.view
+package jp.co.yumemi.android.code_check.ui.compose
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
