+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/compose/LoadingScreen.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.ui.view
+package jp.co.yumemi.android.code_check.ui.compose
 
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.animateFloat
